Refactor: Move Recurrence types in their own public module
