+++ tests/test_dispatch_cli.py
@@ -14,14 +14,14 @@ from tzlocal import get_localzone
 
 from frequenz.client.common.microgrid.components import ComponentCategory
 from frequenz.client.dispatch.__main__ import cli
-from frequenz.client.dispatch.test.client import ALL_KEY, FakeClient
-from frequenz.client.dispatch.types import (
-    Dispatch,
+from frequenz.client.dispatch.recurrence import (
     EndCriteria,
     Frequency,
     RecurrenceRule,
     Weekday,
 )
+from frequenz.client.dispatch.test.client import ALL_KEY, FakeClient
+from frequenz.client.dispatch.types import Dispatch
 
 TEST_NOW = datetime(2023, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
 """Arbitrary time used as NOW for testing."""
